+++ src/components/FormFieldGeldigheid/index.js
@@ -1,0 +1,3 @@
+import FormFieldGeldigheid from "./FormFieldGeldigheid.js"
+
+export default FormFieldGeldigheid
