+++ src/pages/MuteerBeleidsrelatiesCRUD/ContainerCrudFields/ContainerCrudFields.js
@@ -7,7 +7,7 @@ import APIcontext from "./../APIContext"
 import ContainerMain from "./../../../components/ContainerMain"
 import ContainerFormSection from "./../../../components/ContainerFormSection"
 import FormFieldTextArea from "./../../../components/FormFieldTextArea"
-import FormFieldDate from "./../../../components/FormFieldDate"
+import FormFieldGeldigheid from "./../../../components/FormFieldGeldigheid"
 import FormFieldSelectBeleidskeuze from "./../../../components/FormFieldSelectBeleidskeuze"
 
 class ContainerCrudFields extends React.Component {
@@ -71,7 +71,7 @@ class ContainerCrudFields extends React.Component {
                                             {/* Begin Geldigheid */}
                                             {crudObject["Begin_Geldigheid"] !==
                                             undefined ? (
-                                                <FormFieldDate
+                                                <FormFieldGeldigheid
                                                     handleChange={
                                                         this.context
                                                             .handleChange
@@ -93,7 +93,7 @@ class ContainerCrudFields extends React.Component {
                                             {/* Eind Geldigheid */}
                                             {crudObject["Eind_Geldigheid"] !==
                                             undefined ? (
-                                                <FormFieldDate
+                                                <FormFieldGeldigheid
                                                     openUitwerkingstrede={true}
                                                     hideToggleUitwerkingstrede={
                                                         true
