+++ src/pages/MuteerUniversalObjectCRUD/FormFieldContainers/FormFieldContainerBeleidsdoelen/FormFieldContainerBeleidsdoelen.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types"
 
 import ContainerFormSection from "../../../../components/ContainerFormSection"
 import {
-    FormFieldDate,
+    FormFieldGeldigheid,
     FormFieldTextInput,
     FormFieldTextArea,
     FormFieldWeblink,
@@ -58,7 +58,7 @@ function FormFieldContainerBeleidsdoelen({
                 />
 
                 <div className="flex flex-wrap -mx-3">
-                    <FormFieldDate
+                    <FormFieldGeldigheid
                         handleChange={handleChange}
                         fieldValue={crudObject["Begin_Geldigheid"]}
                         fieldLabel="Inwerkingtreding"
@@ -67,7 +67,7 @@ function FormFieldContainerBeleidsdoelen({
                         titleSingular={titleSingular}
                     />
 
-                    <FormFieldDate
+                    <FormFieldGeldigheid
                         openUitwerkingstrede={true}
                         handleChange={handleChange}
                         fieldValue={crudObject["Eind_Geldigheid"]}
