+++ src/pages/MuteerUniversalObjectCRUD/FormFieldContainers/FormFieldContainerBeleidsregels/FormFieldContainerBeleidsregels.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types"
 
 import ContainerFormSection from "./../../../../components/ContainerFormSection"
 import {
-    FormFieldDate,
+    FormFieldGeldigheid,
     FormFieldTextInput,
     FormFieldTextArea,
     FormFieldWeblink,
@@ -67,7 +67,7 @@ function FormFieldContainerBeleidsregels({
                 />
 
                 <div className="flex flex-wrap -mx-3">
-                    <FormFieldDate
+                    <FormFieldGeldigheid
                         handleChange={handleChange}
                         fieldValue={crudObject["Begin_Geldigheid"]}
                         fieldLabel="Inwerkingtreding"
@@ -76,7 +76,7 @@ function FormFieldContainerBeleidsregels({
                         titleSingular={titleSingular}
                     />
 
-                    <FormFieldDate
+                    <FormFieldGeldigheid
                         openUitwerkingstrede={true}
                         handleChange={handleChange}
                         fieldValue={crudObject["Eind_Geldigheid"]}
