+++ src/pages/MuteerUniversalObjectCRUD/FormFieldContainers/FormFieldContainerThemas/FormFieldContainerThemas.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types"
 
 import ContainerFormSection from "./../../../../components/ContainerFormSection"
 import {
-    FormFieldDate,
+    FormFieldGeldigheid,
     FormFieldTextInput,
     FormFieldTextArea,
     FormFieldWeblink,
@@ -53,7 +53,7 @@ function FormFieldContainerThemas({ titleSingular, crudObject, handleChange }) {
                     titleSingular={titleSingular}
                 />
                 <div className="flex flex-wrap -mx-3">
-                    <FormFieldDate
+                    <FormFieldGeldigheid
                         handleChange={handleChange}
                         fieldValue={crudObject["Begin_Geldigheid"]}
                         fieldLabel="Inwerkingtreding"
@@ -61,7 +61,7 @@ function FormFieldContainerThemas({ titleSingular, crudObject, handleChange }) {
                         pValue="Indien bekend, kan hier de datum van inwerkingtreding worden ingevuld"
                         titleSingular={titleSingular}
                     />
-                    <FormFieldDate
+                    <FormFieldGeldigheid
                         openUitwerkingstrede={true}
                         handleChange={handleChange}
                         fieldValue={crudObject["Eind_Geldigheid"]}
